Adapt imports for refactor
