+++ src/lib/dataviews/DataViewWrapper.ts
@@ -1,6 +1,5 @@
 import { WithEvents } from '@/core/mixins/WithEvents';
-import { Layer } from '@/viz';
-import { Source } from '@/viz/source';
+import { Layer, Source } from '@/viz';
 import { Filter } from '@/viz/filters/types';
 import { AggregationType } from '@/data/operations/aggregation/aggregation';
 import { DataViewImpl } from './DataViewImpl';
