+++ src/lib/viz/layer/Layer.ts
@@ -7,7 +7,8 @@ import deepmerge from 'deepmerge';
 import { GeoJSON } from 'geojson';
 import { uuidv4 } from '@/core/utils/uuid';
 import { WithEvents } from '@/core/mixins/WithEvents';
-import { DatasetSource, SQLSource, GeoJSONSource, Source, StatFields } from '@/viz/source';
+import { DatasetSource, SQLSource, GeoJSONSource, Source } from '@/viz';
+import { StatFields } from '@/viz/source';
 import { DOLayer } from '../deck/DOLayer';
 import { getStyles, StyleProperties, Style } from '../style';
 import { ViewportFeaturesGenerator } from '../interactivity/viewport-features/ViewportFeaturesGenerator';
